fix(security): parameterize untrusted SQL inputs across develop (#7210)
* fix(security): parameterize untrusted SQL inputs across develop

Bind, db_qstr, or integer-cast request-derived values that previously
reached SQL by interpolation: RLIKE/REGEXP filters in graph/tree/report
views, IN() id lists in device, profile, and SNMP receiver management,
host_template_id in automation matching, and ORDER BY columns in the
automation match pages (allowlist + ASC/DESC). Validate graph_name_regexp
at save and anchor the graph_template_id filter regex. Covers
CVE-2026-46531.


* style: apply php-cs-fixer single-quote and alignment fixes


* ci: ignore phpstan function.alreadyNarrowedType identifier


---------

+++ lib/html.php
@@ -2433,7 +2433,7 @@ function html_transform_graph_template_ids(mixed $ids) : string {
 		if (is_numeric($id)) {
 			$return_ids[] = $id;
 		} elseif (str_contains($id, 'cg_')) {
-			$new_id       = str_replace('cg_', '', $id);
+			$new_id       = (int) str_replace('cg_', '', $id);
 			$return_ids[] = $new_id;
 		} else {
 			$id = str_replace('dq_', '', $id);
